requirements.txt
# ─── Core ML ───
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0

# ─── API & Frontend ───
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
streamlit>=1.25.0
python-multipart>=0.0.6

# ─── Data Processing ───
h5py>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# ─── Visualization & Evaluation ───
matplotlib>=3.7.0
scipy>=1.11.0

# ─── CLIP (install separately from GitHub) ───
# pip install git+https://github.com/openai/CLIP.git
# OR: pip install clip-by-openai

# ─── Utilities ───
requests>=2.31.0
pydantic>=2.0.0
